+++ src/physiology_analysis_tools/__init__.py
@@ -1,0 +1,3 @@
+import physiology_analysis_tools
+import physiology_analysis_tools.modules as modules
+import physiology_analysis_tools.main as main
